+++ src/azure-cli/azure/cli/command_modules/vm/_actions.py
@@ -191,7 +191,11 @@ def load_images_from_aliases_doc(cli_ctx, publisher=None, offer=None, sku=None, 
         try:
             response = requests.get(target_url, verify=not should_disable_connection_verify())
             if response.status_code == 200:
-                dic = json.loads(response.content.decode())
+                try:
+                    dic = json.loads(response.content.decode())
+                except ValueError:
+                    logger.warning("Failed to parse image alias doc from '%s'. Use local copy instead.", target_url)
+                    dic = json.loads(alias_json)
             else:
                 logger.warning("Failed to retrieve image alias doc '%s'. Error: '%s'. Use local copy instead.",
                                target_url, response)